411. Campground Show
Creat a route for displaying a campground based on the document ObjectId e.g. '/campgrounds/61f80ed...'
Ammend 'index.js' so that the list of campgrounds is a list of links.
Create an ejs file for displaying this single campground called 'show.ejs'
Implement the find within this new route using the route parameter
Display the campground's title, location in 'show.ejs'