487. Intro to Flash
(Back to using Yelpcamp)
The idea of flash to flash up a msg one time in response to something never to be seen again until we redo
that action. In practice it is basically a place in the session to store a flash message. This would be for scenarios like a success, confirmation or failure message that shows up one time and then goes away. So you would display the message in the template only on that occasion it is reqd.
	import 'connect-flash' and use flash middleware (also make sure you have express-session)
	You will also need cookieParser & session middleware I believe to use flash middleware
	To Demo Flash we want to flash up a (single occurence) msg when a CG is added, we don't want to see that msg again:
		In route "app.post('/campgrounds'....":
			Setup a flash message using req.flash() after successfully saving a CG
			Edit temporarily the redirect to instead go to 'res.redirect(`/campgrounds`);'
			In "app.get('/campgrounds'..." update your 'res.render' so as to pass on your flash msg
			In the 'index.ejs' display the flash msg at the top
			Test the index template to see if anything flashes up, there shouldn't be any msg displayed
			Now add a new CG and save. You should now see a list of CGs along with the flash msg at the top
			Now refresh the page and the flash msg should disappear!!